refactor(doubt): move logic to controller and simplify routes
Extracted doubt-related logic from routes into a dedicated controller for better separation of concerns. Simplified routes by removing redundant code and leveraging the new controller functions. This improves maintainability and readability of the codebase.
